+++ tests/test.rs
@@ -2552,6 +2552,30 @@ fn test_xlsx_empty_shared_string() {
     );
 }
 
+// Test for issue #607 where a shared string in a cell doesn't have a <v> index.
+#[test]
+fn test_xlsx_empty_shared_string_value() {
+    let mut excel: Xlsx<_> = wb("empty_shared_string_value.xlsx");
+    let range = excel.worksheet_range("Sheet1").unwrap();
+
+    range_eq!(
+        range,
+        [
+            [
+                String("Name".to_string()),
+                String("City".to_string()),
+                String("Status".to_string())
+            ],
+            [
+                String("Alice".to_string()),
+                Empty,
+                String("Active".to_string())
+            ],
+            [Empty, String("Berlin".to_string()), Empty,]
+        ]
+    );
+}
+
 // Test for issue #587, an xlsx file where table references IDs are absolute
 // instead of relative, i.e., "/xl/tables/table1.xml" instead of the Excel
 // generated "../tables/table1.xml".
@@ -2794,4 +2818,54 @@ fn test_pivot_table_cache_match() {
         .map(|val| val.unwrap())
         .collect::<Vec<_>>();
     assert_eq!(results1, results2);
+}
+
+#[test]
+fn vba_optional_records_581() {
+    // parsing missing-but-optional records should succeed
+    // ref: https://github.com/tafia/calamine/issues/581
+    let mut excel: Xls<_> = wb("optional_records.xls");
+    let vba = excel.vba_project().unwrap().unwrap();
+    let references = vba.get_references();
+    assert_eq!(
+        references,
+        [
+            Reference {
+                name: "stdole".to_string(),
+                description: "OLE Automation".to_string(),
+                path: "C:\\WINNT\\System32\\StdOle2.Tlb".into(),
+            },
+            Reference {
+                name: "MSForms".to_string(),
+                description: "Microsoft Forms 2.0 Object Library".to_string(),
+                path: "C:\\WINNT\\System32\\MSForms.TWD".into(),
+            },
+            Reference {
+                name: "Office".to_string(),
+                description: "Microsoft Office 8.0 Object Library".to_string(),
+                path: "C:\\Program Files\\Microsoft Office\\Office\\MSO97.DLL".into(),
+            },
+        ]
+    );
+    let module_names = vba.get_module_names();
+    assert_eq!(module_names, vec!["Sheet15"]);
+}
+
+#[test]
+fn biff5_defined_names_and_empty_sheets_612() {
+    // covers a parsing error where biff8 byte layout was assumed
+    // ref: https://github.com/tafia/calamine/issues/612
+    let excel: Xls<_> = wb("misc_biff5_parsing.xls");
+    assert_eq!(excel.sheet_names().len(), 11);
+    assert_eq!(
+        excel.defined_names(),
+        [("test".to_string(), "Sheet10!$L$17".to_string())]
+    );
+}
+
+#[test]
+fn test_capitalized_book_stream() {
+    // capitalized WORKBOOK and BOOK stream names are accepted by libreoffice
+    // ref: https://github.com/tafia/calamine/issues/618
+    let _wb: Xls<_> = wb("capitalized_wbook_stream.xls");
 }